Add files via upload
This project aims to define a simplified polymorphic class hierarchy that represents certain types of expressions. The first class to be defined is Expression, which represents mathematical functions of the form y=f(x), where f is a function of one real variable. This class offers four basic functionalities: a function isRoot to test whether a given value of x is a root of the function f, an overloaded function call operator (operator()) to evaluate an expression E given a value d for the variable x, a stream insertion operator operator<< to write an expression to a given output stream, and a function clone supporting polymorphic copying.

The second class is Polynomial, which is a subclass of Expression that represents a polynomial. This class offers additional functionalities, such as a constructor that creates a polynomial from a vector of doubles, a conversion constructor to convert a real constant into a polynomial, a copy constructor, an assignment operator, an add-and-assign operator (operator+=), addition of two polynomials p+q, addition of a polynomial with a real (double) value d (i.e., p+d and d+p), a subscript operator operator[] that can be used to access the value of a polynomial’s coefficient, and the ability to write a polynomial to an output stream.

The third class is Logarithm, which is another subclass of Expression that represents a logarithmic function of the form c1+c2*logb(E), where E is an expression (either a polynomial or a logarithm) and c1, c2, and b are real constants. Logarithm offers all the functionalities of Expression, as well as a default constructor that creates the logarithm log2(x), a constructor that creates a logarithmic expression given an expression E, and constants c1, c2, and b > 1, a copy constructor, an assignment operator, and a function named set_base that modifies the base of the logarithm to a given integer b.